Fix fp16/bf16 dtype mismatch crashing GradScaler on first optimizer step
Model was loaded without torch_dtype, so it defaulted to the
checkpoint's stored bfloat16 for non-quantized params - including the
LoRA adapter weights created afterward by get_peft_model. Training
uses fp16 mixed precision (T4 has no bf16 tensor cores), and torch's
GradScaler.unscale_ has no BFloat16 implementation, so it crashed at
the first clip_grad_norm_ call. Explicit torch_dtype=torch.float16
keeps everything consistent.

+++ src/training/train_qlora.py
@@ -69,7 +69,12 @@ def main() -> None:
         bnb_4bit_use_double_quant=True,
     )
     model = AutoModelForCausalLM.from_pretrained(
-        MODEL_NAME, quantization_config=bnb_config, device_map="auto"
+        MODEL_NAME,
+        quantization_config=bnb_config,
+        device_map="auto",
+        # 명시 안 하면 체크포인트 기본값(bfloat16)을 따라가서, LoRA 어댑터 가중치가
+        # bf16으로 생성되고 fp16 GradScaler가 이를 처리 못해 학습 첫 step에서 터짐.
+        torch_dtype=torch.float16,
     )
     model.config.use_cache = False
     model = prepare_model_for_kbit_training(model)
